panel-scheduled: Change task list order
All panels use the same task list order as task list view.

For scheduled panel it was requested to change the order of the items
and sort them only by due date, in descending order. In order to keep
things a little bit more organized, it was also sorted by title.

A new API was created in task list view to support it and this patch
uses that API to implement the new sort function.

https://bugzilla.gnome.org/show_bug.cgi?id=760846