gtd-task: Add creation date property
It was requested to add another sort by creation time.

gtd-task didn't keep record of that information, so it was necessary to
add a new property to handle that.

EDS prevents creation time field to be updated by the user agent, so it
was added only a method to recover that information and not to set it.

It was also noticed during some tests that some services like Google
don't return a sane creation time. That seems to be a problem between
EDS and Google.

This commit also changes the sorting order:
- Standard Order: complete > priority > due-date > creation-date > title
- Scheduled Panel Order: due-date > completed > priority >
                         creation-date > title

https://bugzilla.gnome.org/show_bug.cgi?id=759878