eds: make it an embedded plugin
EDS is the core plugin for gnome-todo and it should be hidden to
the user. Libpeas added support for embedded plugins (since v0.19)
and this allows EDS to be loaded as an embedded plugin for ToDo.
This should prevent the user from modifying EDS from an install
directory.

EDS resources were added to gresources, to be linked with the
binary executable. Appropriate changes to add resource search path
and compilation with Makefiles were done to allow compatibility.
eds.plugin was added with an 'Embedded' field.

https://bugzilla.gnome.org/show_bug.cgi?id=760847